broke up samples into a package
