+++ autofit/__init__.py
@@ -1,6 +1,7 @@
 from .non_linear.grid.grid_search import GridSearch as SearchGridSearch
 from . import conf
 from . import exc
+from .non_linear.samples import *
 from .database.aggregator import Aggregator
 from .database.aggregator import Query
 from .aggregator.search_output import SearchOutput
@@ -61,11 +62,9 @@ from .non_linear.paths import DatabasePaths
 from .non_linear.paths import make_path
 from .non_linear.result import Result
 from .non_linear.result import ResultsCollection
-from .non_linear.samples import StoredSamples
-from .non_linear.samples import MCMCSamples
-from .non_linear.samples import NestSamples
-from .non_linear.samples import OptimizerSamples
-from .non_linear.samples import PDFSamples
+from .non_linear.samples.stored import StoredSamples
+from .non_linear.samples.mcmc import MCMCSamples
+from .non_linear.samples.nest import NestSamples
 from .mock.mock import Gaussian
 from .text import formatter
 from .text import samples_text
