+++ autofit/non_linear/mcmc/zeus.py
@@ -6,12 +6,12 @@ import zeus
 from sqlalchemy.orm import Session
 
 from autoconf import conf
-from autofit import exc
+from autofit import exc, MCMCSamples
 from autofit.mapper.model_mapper import ModelMapper
 from autofit.mapper.prior_model.abstract import AbstractPriorModel
 from autofit.non_linear.mcmc.abstract_mcmc import AbstractMCMC
 from autofit.non_linear.mcmc.auto_correlations import AutoCorrelationsSettings, AutoCorrelations
-from autofit.non_linear.samples import MCMCSamples, Sample
+from autofit.non_linear.samples import Sample
 from autofit.non_linear.abstract_search import PriorPasser
 from autofit.non_linear.initializer import Initializer
 from autofit.plot import ZeusPlotter
