+++ autofit/non_linear/nest/multi_nest.py
@@ -10,7 +10,8 @@ from autofit.mapper.prior_model.abstract import AbstractPriorModel
 from autofit.non_linear import abstract_search
 from autofit.non_linear import result as res
 from autofit.non_linear.nest import abstract_nest
-from autofit.non_linear.samples import NestSamples, Sample
+from autofit.non_linear.samples import Sample
+from autofit import NestSamples
 from autofit.non_linear.abstract_search import PriorPasser
