--- autofit/non_linear/samples.py
@@ -1,1281 +1,0 @@
-import csv
-import json
-import math
-from copy import copy
-from typing import List, Optional
-
-import numpy as np
-
-from autofit.mapper.model import ModelInstance
-from autofit.mapper.model_mapper import ModelMapper
-from autofit.mapper.prior_model.abstract import AbstractPriorModel
-from autofit.non_linear.mcmc.auto_correlations import AutoCorrelationsSettings
-
-
-def quantile(x, q, weights=None):
-    """
-    Copied from corner.py
-
-    Compute sample quantiles with support for weighted samples.
-
-    Note
-    ----
-    When ``weight_list`` is ``None``, this method simply calls numpy's percentile
-    function with the values of ``q`` multiplied by 100.
-
-    Parameters
-    ----------
-    x : array_like[nsamples,]
-       The samples.
-    q : array_like[nquantiles,]
-       The list of quantiles to compute. These should all be in the range
-       ``[0, 1]``.
-    weights : Optional[array_like[nsamples,]]
-        An optional weight corresponding to each sample. These
-
-    Returns
-    -------
-    quantiles : array_like[nquantiles,]
-        The sample quantiles computed at ``q``.
-
-    Raises
-    ------
-    ValueError
-        For invalid quantiles; ``q`` not in ``[0, 1]`` or dimension mismatch between ``x`` and ``weight_list``.
-    """
-    x = np.atleast_1d(x)
-    q = np.atleast_1d(q)
-
-    if np.any(q < 0.0) or np.any(q > 1.0):
-        raise ValueError("Quantiles must be between 0 and 1")
-
-    if weights is None:
-        return np.percentile(x, list(100.0 * q))
-    else:
-        weights = np.atleast_1d(weights)
-        if len(x) != len(weights):
-            raise ValueError("Dimension mismatch: len(weight_list) != len(x)")
-        idx = np.argsort(x)
-        sw = weights[idx]
-        cdf = np.cumsum(sw)[:-1]
-        cdf /= cdf[-1]
-        cdf = np.append(0, cdf)
-        return np.interp(q, cdf, x[idx]).tolist()
-
-
-class Sample:
-    def __init__(
-            self,
-            log_likelihood: float,
-            log_prior: float,
-            weight: float,
-            kwargs=None
-    ):
-        """
-        One sample taken during a search
-
-        Parameters
-        ----------
-        log_likelihood
-            The likelihood associated with this instance
-        log_prior
-            A logarithmic prior of the instance
-        weight
-            The weight this sample contributes to the PDF.
-        kwargs
-            Dictionary mapping model paths to values for the sample
-        """
-        self.log_likelihood = log_likelihood
-        self.log_prior = log_prior
-        self.weight = weight
-        self.kwargs = kwargs or dict()
-
-    @property
-    def log_posterior(self) -> float:
-        """
-        Compute the posterior
-        """
-        return self.log_likelihood + self.log_prior
-
-    def parameter_lists_for_model(
-            self,
-            model: AbstractPriorModel
-    ) -> List[float]:
-        """
-        Values for instantiating a model, in the same order as priors
-        from the model.
-
-        Parameters
-        ----------
-        model
-            The model from which this was a sample
-
-        Returns
-        -------
-        A list of physical values
-        """
-        if self.is_path_kwargs:
-            paths = model.all_paths
-        else:
-            paths = model.all_names
-
-        return self.parameter_lists_for_paths(
-            paths
-        )
-
-    def parameter_lists_for_paths(
-            self,
-            paths
-    ):
-        result = list()
-        for keys in paths:
-            for key in keys:
-                if key in self.kwargs:
-                    result.append(
-                        self.kwargs[
-                            key
-                        ]
-                    )
-                    break
-                else:
-                    raise KeyError(
-                        f"Could not find any of the following keys in kwargs {keys}"
-                    )
-        return result
-
-    @property
-    def is_path_kwargs(self) -> bool:
-        """
-        Are the keys in the kwargs dictionary tuples? If they
-        are this indicates that they are explicit paths through
-        the model.
-        """
-        for key in self.kwargs.keys():
-            return isinstance(
-                key, tuple
-            )
-        return False
-
-    @classmethod
-    def from_lists(
-            cls,
-            model: AbstractPriorModel,
-            parameter_lists: List[List[float]],
-            log_likelihood_list: List[float],
-            log_prior_list: List[float],
-            weight_list: List[float]
-    ) -> List["Sample"]:
-        """
-        Convenience method to create a list of samples from lists of contained values
-        """
-        samples = list()
-
-        paths = model.unique_prior_paths
-
-        for params, log_likelihood, log_prior, weight in zip(
-                parameter_lists,
-                log_likelihood_list,
-                log_prior_list,
-                weight_list
-        ):
-            arg_dict = {
-                t: param
-                for t, param
-                in zip(
-                    paths,
-                    params
-                )
-            }
-
-            samples.append(
-                cls(
-                    log_likelihood=log_likelihood,
-                    log_prior=log_prior,
-                    weight=weight,
-                    kwargs=arg_dict
-                )
-            )
-        return samples
-
-    def instance_for_model(self, model: AbstractPriorModel):
-        """
-        Create an instance from this sample for a model
-
-        Parameters
-        ----------
-        model
-            The model the this sample was taken from
-
-        Returns
-        -------
-        The instance corresponding to this sample
-        """
-        try:
-            if self.is_path_kwargs:
-                return model.instance_from_path_arguments(
-                    self.kwargs
-                )
-            else:
-                return model.instance_from_prior_name_arguments(
-                    self.kwargs
-                )
-
-        except KeyError:
-            # TODO: Does this get used? If so, why?
-            return model.instance_from_vector(
-                self.parameter_lists_for_model(model)
-            )
-
-
-def load_from_table(filename: str) -> List[Sample]:
-    """
-    Load samples from a table
-
-    Parameters
-    ----------
-    filename
-        The path to a CSV file
-
-    Returns
-    -------
-    A list of samples, one for each row in the CSV
-    """
-    samples = list()
-
-    sample_args = (
-        "log_likelihood",
-        "log_prior",
-        "weight",
-    )
-
-    with open(filename, "r+", newline="") as f:
-        reader = csv.reader(f)
-        headers = next(reader)
-        for row in reader:
-            d = {
-                header: float(value)
-                for header, value
-                in zip(
-                    headers,
-                    row
-                )
-            }
-
-            samples.append(
-                Sample(
-                    **{
-                        key: value
-                        for key, value
-                        in d.items()
-                        if key in sample_args
-                    },
-                    kwargs={
-                        key: value
-                        for key, value
-                        in d.items()
-                        if key not in sample_args
-                    }
-                )
-            )
-
-    return samples
-
-
-class OptimizerSamples:
-    def __init__(
-            self,
-            model: AbstractPriorModel,
-            sample_list: List[Sample],
-            time: Optional[float] = None,
-    ):
-        """
-        The `Samples` of a non-linear search, specifically the samples of an search which only provides
-        information on the global maximum likelihood solutions, but does not map-out the posterior and thus does
-        not provide information on parameter errors.
-
-        Parameters
-        ----------
-        model
-            Maps input vectors of unit parameter values to physical values and model instances via priors.
-        """
-        self.model = model
-        self.sample_list = sample_list
-        self.time = time
-
-        self._paths = None
-        self._names = None
-
-    @property
-    def paths(self):
-        """
-        A list of paths to unique priors in the same order as prior
-        ids (and therefore sample columns)
-
-        Uses hasattr to make backwards compatible
-        """
-        if not hasattr(self, "_paths") or self._paths is None:
-            self._paths = self.model.all_paths
-        return self._paths
-
-    @property
-    def names(self) -> List[str]:
-        """
-        A list of names of unique priors in the same order as prior
-        ids (and therefore sample columns)
-
-        Uses hasattr to make backwards compatible
-        """
-        if not hasattr(self, "_names") or self._names is None:
-            self._names = self.model.all_names
-        return self._names
-
-    @property
-    def parameter_lists(self):
-        result = list()
-        for sample in self.sample_list:
-            tuples = self.paths if sample.is_path_kwargs else self.names
-            result.append(
-                sample.parameter_lists_for_paths(
-                    tuples
-                )
-            )
-
-        return result
-
-    @property
-    def total_samples(self):
-        return len(self.sample_list)
-
-    @property
-    def weight_list(self):
-        return [
-            sample.weight
-            for sample
-            in self.sample_list
-        ]
-
-    @property
-    def log_likelihood_list(self):
-        return [
-            sample.log_likelihood
-            for sample
-            in self.sample_list
-        ]
-
-    @property
-    def log_posterior_list(self):
-        return [
-            sample.log_posterior
-            for sample
-            in self.sample_list
-        ]
-
-    @property
-    def log_prior_list(self):
-        return [
-            sample.log_prior
-            for sample
-            in self.sample_list
-        ]
-
-    @property
-    def parameters_extract(self):
-        return np.asarray(self.parameter_lists).T
-
-    @property
-    def _headers(self) -> List[str]:
-        """
-        Headers for the samples table
-        """
-
-        return self.model.model_component_and_parameter_names + [
-            "log_likelihood",
-            "log_prior",
-            "log_posterior",
-            "weight",
-        ]
-
-    @property
-    def _rows(self) -> List[List[float]]:
-        """
-        Rows in the samples table
-        """
-
-        log_likelihood_list = self.log_likelihood_list
-        log_prior_list = self.log_prior_list
-        log_posterior_list = self.log_posterior_list
-        weight_list = self.weight_list
-
-        for index, row in enumerate(self.parameter_lists):
-            yield row + [
-                log_likelihood_list[index],
-                log_prior_list[index],
-                log_posterior_list[index],
-                weight_list[index],
-            ]
-
-    def write_table(self, filename: str):
-        """
-        Write a table of parameters, posteriors, priors and likelihoods.
-
-        Parameters
-        ----------
-        filename
-            Where the table is to be written
-        """
-
-        with open(filename, "w+", newline="") as f:
-            writer = csv.writer(f)
-            writer.writerow(self._headers)
-            for row in self._rows:
-                writer.writerow(row)
-
-    @property
-    def info_json(self):
-        return {}
-
-    def info_to_json(self, filename):
-        with open(filename, 'w') as outfile:
-            json.dump(self.info_json, outfile)
-
-    @property
-    def max_log_likelihood_sample(self) -> Sample:
-        """
-        The index of the sample with the highest log likelihood.
-        """
-        most_likely_sample = None
-        for sample in self.sample_list:
-            if most_likely_sample is None or sample.log_likelihood > most_likely_sample.log_likelihood:
-                most_likely_sample = sample
-        return most_likely_sample
-
-    @property
-    def max_log_posterior_sample(self) -> Sample:
-        return self.sample_list[
-            self.max_log_posterior_index
-        ]
-
-    @property
-    def max_log_likelihood_vector(self) -> [float]:
-        """
-        The parameters of the maximum log likelihood sample of the `NonLinearSearch` returned as a list of values.
-        """
-        sample = self.max_log_likelihood_sample
-        return sample.parameter_lists_for_paths(
-            self.paths if sample.is_path_kwargs else self.names
-        )
-
-    @property
-    def max_log_likelihood_instance(self) -> ModelInstance:
-        """
-        The parameters of the maximum log likelihood sample of the `NonLinearSearch` returned as a model instance.
-        """
-        return self.max_log_likelihood_sample.instance_for_model(
-            self.model
-        )
-
-    @property
-    def max_log_posterior_index(self) -> int:
-        """
-        The index of the sample with the highest log posterior.
-        """
-        return int(np.argmax(self.log_posterior_list))
-
-    @property
-    def max_log_posterior_vector(self) -> [float]:
-        """
-        The parameters of the maximum log posterior sample of the `NonLinearSearch` returned as a list of values.
-        """
-        return self.parameter_lists[self.max_log_posterior_index]
-
-    @property
-    def max_log_posterior_instance(self) -> ModelInstance:
-        """
-        The parameters of the maximum log posterior sample of the `NonLinearSearch` returned as a model instance.
-        """
-        return self.model.instance_from_vector(vector=self.max_log_posterior_vector)
-
-    def gaussian_priors_at_sigma(self, sigma: float) -> [list]:
-        """
-        `GaussianPrior`s of every parameter used to link its inferred values and errors to priors used to sample the
-        same (or similar) parameters in a subsequent search, where:
-
-        - The mean is given by maximum log likelihood model values.
-        - Their errors are omitted, as this information is not available from an search. When these priors are
-        used to link to another search, it will thus automatically use the prior config values.
-
-        Parameters
-        -----------
-        sigma
-            The sigma limit within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
-        """
-        return list(map(lambda vector: (vector, 0.0), self.max_log_likelihood_vector))
-
-    def instance_from_sample_index(self, sample_index) -> ModelInstance:
-        """
-        The parameters of an individual sample of the non-linear search, returned as a model instance.
-
-        Parameters
-        -----------
-        sample_index : int
-            The sample index of the weighted sample to return.
-        """
-        return self.model.instance_from_vector(vector=self.parameter_lists[sample_index])
-
-    def minimise(self) -> "OptimizerSamples":
-        """
-        A copy of this object with only important samples retained
-        """
-        samples = copy(self)
-        samples.model = None
-        samples.sample_list = list({
-            self.max_log_likelihood_sample,
-            self.max_log_posterior_sample
-        })
-        return samples
-
-
-class PDFSamples(OptimizerSamples):
-    def __init__(
-            self,
-            model: AbstractPriorModel,
-            sample_list: List[Sample],
-            unconverged_sample_size: int = 100,
-            time: Optional[float] = None,
-    ):
-        """
-        The `Samples` of a non-linear search, specifically the samples of a `NonLinearSearch` which maps out the
-        posterior of parameter space and thus does provide information on parameter errors.
-
-        Parameters
-        ----------
-        model
-            Maps input vectors of unit parameter values to physical values and model instances via priors.
-        """
-
-        super().__init__(
-            model=model,
-            sample_list=sample_list,
-            time=time,
-        )
-
-        self._unconverged_sample_size = int(unconverged_sample_size)
-
-    @classmethod
-    def from_table(self, filename: str, model, number_live_points=None):
-        """
-        Write a table of parameters, posteriors, priors and likelihoods
-
-        Parameters
-        ----------
-        filename
-            Where the table is to be written
-        """
-
-        sample_list = load_from_table(
-            filename=filename
-        )
-
-        return StoredSamples(
-            model=model,
-            sample_list=sample_list
-        )
-
-    @property
-    def unconverged_sample_size(self):
-        """
-        If a set of samples are unconverged, alternative methods to compute their means, errors, etc are used as
-        an alternative to corner.py.
-
-        These use a subset of samples spanning the range from the most recent sample to the valaue of the
-        unconverted_sample_size. However, if there are fewer samples than this size, we change the size to be the
-        the size of the total number of samples.
-        """
-        if self.total_samples > self._unconverged_sample_size:
-            return self._unconverged_sample_size
-        return self.total_samples
-
-    @property
-    def pdf_converged(self) -> bool:
-        """
-        To analyse and visualize samples the analysis must be sufficiently converged to produce smooth enough
-        PDF for error estimate and PDF generation.
-
-        This property checks whether the non-linear search's samples are sufficiently converged for this, by checking
-        if one sample's weight contains > 99% of the weight. If this is the case, it implies the convergence necessary
-        for error estimate and visualization has not been met.
-
-        This does not necessarily imply the `NonLinearSearch` has converged overall, only that errors and visualization
-        can be performed numerically.
-        """
-        if np.max(self.weight_list) > 0.99:
-            return False
-        return True
-
-    @property
-    def median_pdf_vector(self) -> [float]:
-        """
-        The median of the probability density function (PDF) of every parameter marginalized in 1D, returned
-        as a list of values.
-        """
-        if self.pdf_converged:
-            return [
-                quantile(x=params, q=0.5, weights=self.weight_list)[0]
-                for params in self.parameters_extract
-            ]
-        return self.max_log_likelihood_vector
-
-    @property
-    def median_pdf_instance(self) -> ModelInstance:
-        """
-        The median of the probability density function (PDF) of every parameter marginalized in 1D, returned
-        as a model instance.
-        """
-        return self.model.instance_from_vector(vector=self.median_pdf_vector)
-
-    def vector_at_sigma(self, sigma: float) -> [(float, float)]:
-        """
-        The value of every parameter marginalized in 1D at an input sigma value of its probability density function
-        (PDF), returned as two lists of values corresponding to the lower and upper values parameter values.
-
-        For example, if sigma is 1.0, the marginalized values of every parameter at 31.7% and 68.2% percentiles of each
-        PDF is returned.
-
-        This does not account for covariance between parameters. For example, if two parameters (x, y) are degenerate
-        whereby x decreases as y gets larger to give the same PDF, this function will still return both at their
-        upper values. Thus, caution is advised when using the function to reperform a model-fits.
-
-        For Dynesty, this is estimated using corner.py if the samples have converged, by sampling the density
-        function at an input PDF %. If not converged, a crude estimate using the range of values of the current
-        physical live points is used.
-
-        Parameters
-        ----------
-        sigma
-            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
-        """
-
-        if self.pdf_converged:
-
-            low_limit = (1 - math.erf(sigma / math.sqrt(2))) / 2
-
-            lower_errors = [
-                quantile(x=params, q=low_limit, weights=self.weight_list)[0]
-                for params in self.parameters_extract
-            ]
-            upper_errors = [
-                quantile(x=params, q=1 - low_limit, weights=self.weight_list)[0]
-                for params in self.parameters_extract
-            ]
-
-            return [(lower, upper) for lower, upper in zip(lower_errors, upper_errors)]
-
-        parameters_min = list(
-            np.min(self.parameter_lists[-self.unconverged_sample_size:], axis=0)
-        )
-        parameters_max = list(
-            np.max(self.parameter_lists[-self.unconverged_sample_size:], axis=0)
-        )
-
-        return [
-            (parameters_min[index], parameters_max[index])
-            for index in range(len(parameters_min))
-        ]
-
-    def vector_at_upper_sigma(self, sigma: float) -> [float]:
-        """
-        The upper value of every parameter marginalized in 1D at an input sigma value of its probability density
-        function (PDF), returned as a list.
-
-        See vector_at_sigma for a full description of how the parameters at sigma are computed.
-
-        Parameters
-        -----------
-        sigma
-            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
-        """
-        return list(map(lambda param: param[1], self.vector_at_sigma(sigma)))
-
-    def vector_at_lower_sigma(self, sigma: float) -> [float]:
-        """
-        The lower value of every parameter marginalized in 1D at an input sigma value of its probability density
-        function (PDF), returned as a list.
-
-        See vector_at_sigma for a full description of how the parameters at sigma are computed.
-
-        Parameters
-        -----------
-        sigma
-            The sigma limit within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
-        """
-        return list(map(lambda param: param[0], self.vector_at_sigma(sigma)))
-
-    def instance_at_sigma(self, sigma: float) -> ModelInstance:
-        """
-        The value of every parameter marginalized in 1D at an input sigma value of its probability density function
-        (PDF), returned as a list of model instances corresponding to the lower and upper values estimated from the PDF.
-
-        See vector_at_sigma for a full description of how the parameters at sigma are computed.
-
-        Parameters
-        ----------
-        sigma
-            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
-        """
-        return self.model.instance_from_vector(
-            vector=self.vector_at_sigma(sigma=sigma), assert_priors_in_limits=False
-        )
-
-    def instance_at_upper_sigma(self, sigma: float) -> ModelInstance:
-        """
-        The upper value of every parameter marginalized in 1D at an input sigma value of its probability density
-        function (PDF), returned as a model instance.
-
-        See vector_at_sigma for a full description of how the parameters at sigma are computed.
-
-        Parameters
-        -----------
-        sigma
-            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
-        """
-        return self.model.instance_from_vector(
-            vector=self.vector_at_upper_sigma(sigma=sigma),
-            assert_priors_in_limits=False,
-        )
-
-    def instance_at_lower_sigma(self, sigma: float) -> ModelInstance:
-        """
-        The lower value of every parameter marginalized in 1D at an input sigma value of its probability density
-        function (PDF), returned as a model instance.
-
-        See vector_at_sigma for a full description of how the parameters at sigma are computed.
-
-        Parameters
-        -----------
-        sigma
-            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
-        """
-        return self.model.instance_from_vector(
-            vector=self.vector_at_lower_sigma(sigma=sigma),
-            assert_priors_in_limits=False,
-        )
-
-    def error_vector_at_sigma(self, sigma: float) -> [(float, float)]:
-        """
-        The lower and upper error of every parameter marginalized in 1D at an input sigma value of its probability
-        density function (PDF), returned as a list.
-
-        See vector_at_sigma for a full description of how the parameters at sigma are computed.
-
-        Parameters
-        -----------
-        sigma
-            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
-        """
-        error_vector_lower = self.error_vector_at_lower_sigma(sigma=sigma)
-        error_vector_upper = self.error_vector_at_upper_sigma(sigma=sigma)
-        return [(lower, upper) for lower, upper in zip(error_vector_lower, error_vector_upper)]
-
-    def error_vector_at_upper_sigma(self, sigma: float) -> [float]:
-        """
-        The upper error of every parameter marginalized in 1D at an input sigma value of its probability density
-        function (PDF), returned as a list.
-
-        See vector_at_sigma for a full description of how the parameters at sigma are computed.
-
-        Parameters
-        -----------
-        sigma
-            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
-        """
-        uppers = self.vector_at_upper_sigma(sigma=sigma)
-        return list(
-            map(
-                lambda upper, median_pdf: upper - median_pdf,
-                uppers,
-                self.median_pdf_vector,
-            )
-        )
-
-    def error_vector_at_lower_sigma(self, sigma: float) -> [float]:
-        """
-        The lower error of every parameter marginalized in 1D at an input sigma value of its probability density
-        function (PDF), returned as a list.
-
-        See vector_at_sigma for a full description of how the parameters at sigma are computed.
-
-        Parameters
-        -----------
-        sigma
-            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
-        """
-        lowers = self.vector_at_lower_sigma(sigma=sigma)
-        return list(
-            map(
-                lambda lower, median_pdf: median_pdf - lower,
-                lowers,
-                self.median_pdf_vector,
-            )
-        )
-
-    def error_magnitude_vector_at_sigma(self, sigma: float) -> [float]:
-        """
-        The magnitude of every error after marginalization in 1D at an input sigma value of the probability density
-        function (PDF), returned as two lists of values corresponding to the lower and upper errors.
-
-        For example, if sigma is 1.0, the difference in the inferred values marginalized at 31.7% and 68.2% percentiles
-        of each PDF is returned.
-
-        Parameters
-        ----------
-        sigma
-            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
-        """
-        uppers = self.vector_at_upper_sigma(sigma=sigma)
-        lowers = self.vector_at_lower_sigma(sigma=sigma)
-        return list(map(lambda upper, lower: upper - lower, uppers, lowers))
-
-    def error_instance_at_sigma(self, sigma: float) -> ModelInstance:
-        """
-        The error of every parameter marginalized in 1D at an input sigma value of its probability density function
-        (PDF), returned as a list of model instances corresponding to the lower and upper errors.
-
-        See vector_at_sigma for a full description of how the parameters at sigma are computed.
-
-        Parameters
-        ----------
-        sigma
-            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
-        """
-        return self.model.instance_from_vector(
-            vector=self.error_magnitude_vector_at_sigma(sigma=sigma),
-            assert_priors_in_limits=False,
-        )
-
-    def error_instance_at_upper_sigma(self, sigma: float) -> ModelInstance:
-        """
-        The upper error of every parameter marginalized in 1D at an input sigma value of its probability density
-        function (PDF), returned as a model instance.
-
-        See vector_at_sigma for a full description of how the parameters at sigma are computed.
-
-        Parameters
-        -----------
-        sigma
-            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
-        """
-        return self.model.instance_from_vector(
-            vector=self.error_vector_at_upper_sigma(sigma=sigma),
-            assert_priors_in_limits=False,
-        )
-
-    def error_instance_at_lower_sigma(self, sigma: float) -> ModelInstance:
-        """
-        The lower error of every parameter marginalized in 1D at an input sigma value of its probability density
-        function (PDF), returned as a model instance.
-
-        See vector_at_sigma for a full description of how the parameters at sigma are computed.
-
-        Parameters
-        -----------
-        sigma
-            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
-        """
-        return self.model.instance_from_vector(
-            vector=self.error_vector_at_lower_sigma(sigma=sigma),
-            assert_priors_in_limits=False,
-        )
-
-    def gaussian_priors_at_sigma(self, sigma: float) -> [list]:
-        """
-        `GaussianPrior`s of every parameter used to link its inferred values and errors to priors used to sample the
-        same (or similar) parameters in a subsequent search, where:
-
-        - The mean is given by their most-probable values (using median_pdf_vector).
-        - Their errors are computed at an input sigma value (using errors_at_sigma).
-
-        Parameters
-        -----------
-        sigma
-            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
-        """
-
-        means = self.median_pdf_vector
-
-        uppers = self.vector_at_upper_sigma(sigma=sigma)
-        lowers = self.vector_at_lower_sigma(sigma=sigma)
-
-        # noinspection PyArgumentList
-        sigmas = list(
-            map(
-                lambda mean, upper, lower: max([upper - mean, mean - lower]),
-                means,
-                uppers,
-                lowers,
-            )
-        )
-
-        return list(map(lambda mean, sigma: (mean, sigma), means, sigmas))
-
-    def log_likelihood_from_sample_index(self, sample_index: int) -> float:
-        """
-        The log likelihood of an individual sample of the non-linear search.
-
-        Parameters
-        ----------
-        sample_index
-            The index of the sample in the non-linear search, e.g. 0 gives the first sample.
-        """
-        raise NotImplementedError()
-
-    def vector_from_sample_index(self, sample_index: int) -> [float]:
-        """
-        The parameters of an individual sample of the non-linear search, returned as a 1D list.
-
-        Parameters
-        ----------
-        sample_index
-            The index of the sample in the non-linear search, e.g. 0 gives the first sample.
-        """
-        raise NotImplementedError()
-
-    def offset_vector_from_input_vector(self, input_vector: List) -> [float]:
-        """
-        The values of an input_vector offset by the median_pdf_vector (the PDF medians).
-
-        If the 'true' values of a model are known and input as the input_vector, this function returns the results
-        of the `NonLinearSearch` as values offset from the 'true' model. For example, a value 0.0 means the non-linear
-        search estimated the model parameter value correctly.
-
-        Parameters
-        ----------
-        input_vector
-            A 1D list of values the most-probable model is offset by. This list must have the same dimensions as free
-            parameters in the model-fit.
-        """
-        return list(
-            map(
-                lambda input, median_pdf: median_pdf - input,
-                input_vector,
-                self.median_pdf_vector,
-            )
-        )
-
-
-class MCMCSamples(PDFSamples):
-
-    def __init__(
-            self,
-            model: ModelMapper,
-            sample_list: List[Sample],
-            auto_correlation_settings: AutoCorrelationsSettings,
-            unconverged_sample_size: int = 100,
-            time: Optional[float] = None,
-    ):
-
-        self.auto_correlation_settings = auto_correlation_settings
-
-        super().__init__(
-            model=model,
-            sample_list=sample_list,
-            unconverged_sample_size=unconverged_sample_size,
-            time=time,
-        )
-
-    @property
-    def total_walkers(self):
-        raise NotImplementedError
-
-    @property
-    def total_steps(self):
-        raise NotImplementedError
-
-    @property
-    def auto_correlations(self):
-        raise NotImplementedError
-
-    @classmethod
-    def from_table(self, filename: str, model: ModelMapper, number_live_points: int = None):
-        """
-        Write a table of parameters, posteriors, priors and likelihoods
-
-        Parameters
-        ----------
-        filename
-            Where the table is to be written
-        """
-
-        sample_list = load_from_table(filename=filename)
-
-        return OptimizerSamples(
-            model=model,
-            sample_list=sample_list
-        )
-
-    @property
-    def info_json(self):
-        return {
-            "times": None,
-            "check_size": self.auto_correlations.check_size,
-            "required_length": self.auto_correlations.required_length,
-            "change_threshold": self.auto_correlations.change_threshold,
-            "total_walkers": self.total_walkers,
-            "total_steps": self.total_steps,
-            "time": self.time,
-        }
-
-    @property
-    def pdf_converged(self):
-        """
-        To analyse and visualize samples using corner.py, the analysis must be sufficiently converged to produce
-        smooth enough PDF for analysis. This property checks whether the non-linear search's samples are sufficiently
-        converged for corner.py use.
-
-        Emcee samples can be analysed by corner.py irrespective of how long the sampler has run, albeit low run times
-        will likely produce inaccurate results.
-        """
-        try:
-            samples_after_burn_in = self.samples_after_burn_in
-            if len(samples_after_burn_in) == 0:
-                return False
-            return True
-        except ValueError:
-            return False
-
-    @property
-    def samples_after_burn_in(self) -> [list]:
-        """
-        The emcee samples with the initial burn-in samples removed.
-
-        The burn-in period is estimated using the auto-correlation times of the parameters.
-        """
-        raise NotImplementedError()
-
-    @property
-    def converged(self) -> bool:
-        """
-        Whether the emcee samples have converged on a solution or if they are still in a burn-in period, based on the
-        auto correlation times of parameters.
-        """
-        return self.auto_correlations.check_if_converged(
-            total_samples=self.total_samples
-        )
-
-    @property
-    def median_pdf_vector(self) -> [float]:
-        """
-        The median of the probability density function (PDF) of every parameter marginalized in 1D, returned
-        as a list of values.
-
-        This is computed by binning all sampls after burn-in into a histogram and take its median (e.g. 50%) value.
-        """
-
-        if self.pdf_converged:
-            return [
-                float(np.percentile(self.samples_after_burn_in[:, i], [50]))
-                for i in range(self.model.prior_count)
-            ]
-
-        return self.max_log_likelihood_vector
-
-    def vector_at_sigma(self, sigma: float) -> [float]:
-        """
-        The value of every parameter marginalized in 1D at an input sigma value of its probability density function
-        (PDF), returned as two lists of values corresponding to the lower and upper values parameter values.
-
-        For example, if sigma is 1.0, the marginalized values of every parameter at 31.7% and 68.2% percentiles of each
-        PDF is returned.
-
-        This does not account for covariance between parameters. For example, if two parameters (x, y) are degenerate
-        whereby x decreases as y gets larger to give the same PDF, this function will still return both at their
-        upper values. Thus, caution is advised when using the function to reperform a model-fits.
-
-        For Emcee, if the samples have converged this is estimated by binning the samples after burn-in into a
-        histogram and taking the parameter values at the input PDF %.
-
-        Parameters
-        ----------
-        sigma
-            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
-        """
-        limit = math.erf(0.5 * sigma * math.sqrt(2))
-
-        if self.pdf_converged:
-            samples = self.samples_after_burn_in
-
-            return [
-                tuple(
-                    np.percentile(samples[:, i], [100.0 * (1.0 - limit), 100.0 * limit])
-                )
-                for i in range(self.model.prior_count)
-            ]
-
-        parameters_min = list(
-            np.min(self.parameter_lists[-self.unconverged_sample_size:], axis=0)
-        )
-        parameters_max = list(
-            np.max(self.parameter_lists[-self.unconverged_sample_size:], axis=0)
-        )
-
-        return [
-            (parameters_min[index], parameters_max[index])
-            for index in range(len(parameters_min))
-        ]
-
-    @property
-    def log_evidence(self):
-        return None
-
-
-class NestSamples(PDFSamples):
-    def __init__(
-            self,
-            model: AbstractPriorModel,
-            sample_list: List[Sample],
-            unconverged_sample_size: int = 100,
-            time: Optional[float] = None,
-    ):
-        """
-        The Output classes in PyAutoFit provide an interface between the results of a `NonLinearSearch`
-        (e.g.as files on your hard-disk) and Python.
-
-        For example, the output class can be used to load an instance of the best-fit model, get an instance of any
-        individual sample by the `NonLinearSearch` and return information on the likelihoods, errors, etc.
-
-        The Bayesian log evidence estimated by the nested sampling algorithm.
-
-        Parameters
-        ----------
-        model
-            Maps input vectors of unit parameter values to physical values and model instances via priors.
-        """
-
-        super().__init__(
-            model=model,
-            sample_list=sample_list,
-            unconverged_sample_size=unconverged_sample_size,
-            time=time,
-        )
-
-    @property
-    def number_live_points(self):
-        raise NotImplementedError
-
-    @property
-    def log_evidence(self):
-        raise NotImplementedError
-
-    @property
-    def total_samples(self):
-        raise NotImplementedError
-
-    @property
-    def info_json(self):
-        return {
-            "log_evidence": self.log_evidence,
-            "total_samples": self.total_samples,
-            "unconverged_sample_size": self.unconverged_sample_size,
-            "time": self.time,
-            "number_live_points": self.number_live_points
-        }
-
-    @property
-    def total_accepted_samples(self) -> int:
-        """
-        The total number of accepted samples performed by the nested sampler.
-        """
-        return len(self.log_likelihood_list)
-
-    @property
-    def acceptance_ratio(self) -> float:
-        """
-        The ratio of accepted samples to total samples.
-        """
-        return self.total_accepted_samples / self.total_samples
-
-    def samples_within_parameter_range(
-            self,
-            parameter_index: int,
-            parameter_range: [float, float]
-    ) -> "StoredSamples":
-        """
-        Returns a new set of Samples where all points without parameter values inside a specified range removed.
-
-        For example, if our `Samples` object was for a model with 4 parameters are consistent of the following 3 sets
-        of parameters:
-
-        [[1.0, 2.0, 3.0, 4.0]]
-        [[100.0, 2.0, 3.0, 4.0]]
-        [[1.0, 2.0, 3.0, 4.0]]
-
-        This function for `parameter_index=0` and `parameter_range=[0.0, 99.0]` would remove the second sample because
-        the value 100.0 is outside the range 0.0 -> 99.0.
-
-        Parameters
-        ----------
-        parameter_index
-            The 1D index of the parameter (in the model's vector representation) whose values lie between the parameter
-            range if a sample is kept.
-        parameter_range
-            The minimum and maximum values of the range of parameter values this parameter must lie between for it
-            to be kept.
-        """
-
-        parameter_list = []
-        log_likelihood_list = []
-        log_prior_list = []
-        weight_list = []
-
-        for sample in self.sample_list:
-
-            parameters = sample.parameter_lists_for_model(model=self.model)
-
-            if (parameters[parameter_index] > parameter_range[0]) and (
-                    parameters[parameter_index] < parameter_range[1]):
-
-                parameter_list.append(parameters)
-                log_likelihood_list.append(sample.log_likelihood)
-                log_prior_list.append(sample.log_prior)
-                weight_list.append(sample.weight)
-
-        sample_list = Sample.from_lists(
-            model=self.model,
-            parameter_lists=parameter_list,
-            log_likelihood_list=log_likelihood_list,
-            log_prior_list=log_prior_list,
-            weight_list=weight_list
-        )
-
-        return StoredSamples(
-            model=self.model,
-            sample_list=sample_list,
-            unconverged_sample_size=self.unconverged_sample_size,
-        )
-
-
-class StoredSamples(PDFSamples):
-
-    def __init__(
-            self,
-            model: AbstractPriorModel,
-            sample_list: List[Sample],
-            unconverged_sample_size: int = 100,
-            time: Optional[float] = None,
-    ):
-        """
-        The `Samples` of a non-linear search, specifically the samples of a `NonLinearSearch` which maps out the
-        posterior of parameter space and thus does provide information on parameter errors.
-
-        Parameters
-        ----------
-        model : af.ModelMapper
-            Maps input vectors of unit parameter values to physical values and model instances via priors.
-        """
-
-        super().__init__(
-            model=model,
-            sample_list=sample_list,
-            time=time,
-        )
-
-        self._unconverged_sample_size = int(unconverged_sample_size)
