+++ autofit/non_linear/samples/__init__.py
@@ -1,0 +1,6 @@
+from .mcmc import MCMCSamples
+from .nest import NestSamples
+from .optimizer import OptimizerSamples
+from .pdf import PDFSamples
+from .sample import Sample, load_from_table
+from .stored import StoredSamples
