+++ autofit/non_linear/samples/pdf.py
@@ -1,0 +1,488 @@
+import math
+from typing import List, Optional
+
+import numpy as np
+
+from autofit.mapper.model import ModelInstance
+from autofit.mapper.prior_model.abstract import AbstractPriorModel
+from autofit.non_linear.samples.sample import Sample, load_from_table
+from .optimizer import OptimizerSamples
+
+
+class PDFSamples(OptimizerSamples):
+    def __init__(
+            self,
+            model: AbstractPriorModel,
+            sample_list: List[Sample],
+            unconverged_sample_size: int = 100,
+            time: Optional[float] = None,
+    ):
+        """
+        The `Samples` of a non-linear search, specifically the samples of a `NonLinearSearch` which maps out the
+        posterior of parameter space and thus does provide information on parameter errors.
+
+        Parameters
+        ----------
+        model
+            Maps input vectors of unit parameter values to physical values and model instances via priors.
+        """
+
+        super().__init__(
+            model=model,
+            sample_list=sample_list,
+            time=time,
+        )
+
+        self._unconverged_sample_size = int(unconverged_sample_size)
+
+    @classmethod
+    def from_table(cls, filename: str, model, number_live_points=None):
+        """
+        Write a table of parameters, posteriors, priors and likelihoods
+
+        Parameters
+        ----------
+        filename
+            Where the table is to be written
+        """
+        from .stored import StoredSamples
+
+        sample_list = load_from_table(
+            filename=filename
+        )
+
+        return StoredSamples(
+            model=model,
+            sample_list=sample_list
+        )
+
+    @property
+    def unconverged_sample_size(self):
+        """
+        If a set of samples are unconverged, alternative methods to compute their means, errors, etc are used as
+        an alternative to corner.py.
+
+        These use a subset of samples spanning the range from the most recent sample to the valaue of the
+        unconverted_sample_size. However, if there are fewer samples than this size, we change the size to be the
+        the size of the total number of samples.
+        """
+        if self.total_samples > self._unconverged_sample_size:
+            return self._unconverged_sample_size
+        return self.total_samples
+
+    @property
+    def pdf_converged(self) -> bool:
+        """
+        To analyse and visualize samples the analysis must be sufficiently converged to produce smooth enough
+        PDF for error estimate and PDF generation.
+
+        This property checks whether the non-linear search's samples are sufficiently converged for this, by checking
+        if one sample's weight contains > 99% of the weight. If this is the case, it implies the convergence necessary
+        for error estimate and visualization has not been met.
+
+        This does not necessarily imply the `NonLinearSearch` has converged overall, only that errors and visualization
+        can be performed numerically.
+        """
+        if np.max(self.weight_list) > 0.99:
+            return False
+        return True
+
+    @property
+    def median_pdf_vector(self) -> [float]:
+        """
+        The median of the probability density function (PDF) of every parameter marginalized in 1D, returned
+        as a list of values.
+        """
+        if self.pdf_converged:
+            return [
+                quantile(x=params, q=0.5, weights=self.weight_list)[0]
+                for params in self.parameters_extract
+            ]
+        return self.max_log_likelihood_vector
+
+    @property
+    def median_pdf_instance(self) -> ModelInstance:
+        """
+        The median of the probability density function (PDF) of every parameter marginalized in 1D, returned
+        as a model instance.
+        """
+        return self.model.instance_from_vector(vector=self.median_pdf_vector)
+
+    def vector_at_sigma(self, sigma: float) -> [(float, float)]:
+        """
+        The value of every parameter marginalized in 1D at an input sigma value of its probability density function
+        (PDF), returned as two lists of values corresponding to the lower and upper values parameter values.
+
+        For example, if sigma is 1.0, the marginalized values of every parameter at 31.7% and 68.2% percentiles of each
+        PDF is returned.
+
+        This does not account for covariance between parameters. For example, if two parameters (x, y) are degenerate
+        whereby x decreases as y gets larger to give the same PDF, this function will still return both at their
+        upper values. Thus, caution is advised when using the function to reperform a model-fits.
+
+        For Dynesty, this is estimated using corner.py if the samples have converged, by sampling the density
+        function at an input PDF %. If not converged, a crude estimate using the range of values of the current
+        physical live points is used.
+
+        Parameters
+        ----------
+        sigma
+            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
+        """
+
+        if self.pdf_converged:
+
+            low_limit = (1 - math.erf(sigma / math.sqrt(2))) / 2
+
+            lower_errors = [
+                quantile(x=params, q=low_limit, weights=self.weight_list)[0]
+                for params in self.parameters_extract
+            ]
+            upper_errors = [
+                quantile(x=params, q=1 - low_limit, weights=self.weight_list)[0]
+                for params in self.parameters_extract
+            ]
+
+            return [(lower, upper) for lower, upper in zip(lower_errors, upper_errors)]
+
+        parameters_min = list(
+            np.min(self.parameter_lists[-self.unconverged_sample_size:], axis=0)
+        )
+        parameters_max = list(
+            np.max(self.parameter_lists[-self.unconverged_sample_size:], axis=0)
+        )
+
+        return [
+            (parameters_min[index], parameters_max[index])
+            for index in range(len(parameters_min))
+        ]
+
+    def vector_at_upper_sigma(self, sigma: float) -> [float]:
+        """
+        The upper value of every parameter marginalized in 1D at an input sigma value of its probability density
+        function (PDF), returned as a list.
+
+        See vector_at_sigma for a full description of how the parameters at sigma are computed.
+
+        Parameters
+        -----------
+        sigma
+            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
+        """
+        return list(map(lambda param: param[1], self.vector_at_sigma(sigma)))
+
+    def vector_at_lower_sigma(self, sigma: float) -> [float]:
+        """
+        The lower value of every parameter marginalized in 1D at an input sigma value of its probability density
+        function (PDF), returned as a list.
+
+        See vector_at_sigma for a full description of how the parameters at sigma are computed.
+
+        Parameters
+        -----------
+        sigma
+            The sigma limit within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
+        """
+        return list(map(lambda param: param[0], self.vector_at_sigma(sigma)))
+
+    def instance_at_sigma(self, sigma: float) -> ModelInstance:
+        """
+        The value of every parameter marginalized in 1D at an input sigma value of its probability density function
+        (PDF), returned as a list of model instances corresponding to the lower and upper values estimated from the PDF.
+
+        See vector_at_sigma for a full description of how the parameters at sigma are computed.
+
+        Parameters
+        ----------
+        sigma
+            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
+        """
+        return self.model.instance_from_vector(
+            vector=self.vector_at_sigma(sigma=sigma), assert_priors_in_limits=False
+        )
+
+    def instance_at_upper_sigma(self, sigma: float) -> ModelInstance:
+        """
+        The upper value of every parameter marginalized in 1D at an input sigma value of its probability density
+        function (PDF), returned as a model instance.
+
+        See vector_at_sigma for a full description of how the parameters at sigma are computed.
+
+        Parameters
+        -----------
+        sigma
+            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
+        """
+        return self.model.instance_from_vector(
+            vector=self.vector_at_upper_sigma(sigma=sigma),
+            assert_priors_in_limits=False,
+        )
+
+    def instance_at_lower_sigma(self, sigma: float) -> ModelInstance:
+        """
+        The lower value of every parameter marginalized in 1D at an input sigma value of its probability density
+        function (PDF), returned as a model instance.
+
+        See vector_at_sigma for a full description of how the parameters at sigma are computed.
+
+        Parameters
+        -----------
+        sigma
+            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
+        """
+        return self.model.instance_from_vector(
+            vector=self.vector_at_lower_sigma(sigma=sigma),
+            assert_priors_in_limits=False,
+        )
+
+    def error_vector_at_sigma(self, sigma: float) -> [(float, float)]:
+        """
+        The lower and upper error of every parameter marginalized in 1D at an input sigma value of its probability
+        density function (PDF), returned as a list.
+
+        See vector_at_sigma for a full description of how the parameters at sigma are computed.
+
+        Parameters
+        -----------
+        sigma
+            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
+        """
+        error_vector_lower = self.error_vector_at_lower_sigma(sigma=sigma)
+        error_vector_upper = self.error_vector_at_upper_sigma(sigma=sigma)
+        return [(lower, upper) for lower, upper in zip(error_vector_lower, error_vector_upper)]
+
+    def error_vector_at_upper_sigma(self, sigma: float) -> [float]:
+        """
+        The upper error of every parameter marginalized in 1D at an input sigma value of its probability density
+        function (PDF), returned as a list.
+
+        See vector_at_sigma for a full description of how the parameters at sigma are computed.
+
+        Parameters
+        -----------
+        sigma
+            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
+        """
+        uppers = self.vector_at_upper_sigma(sigma=sigma)
+        return list(
+            map(
+                lambda upper, median_pdf: upper - median_pdf,
+                uppers,
+                self.median_pdf_vector,
+            )
+        )
+
+    def error_vector_at_lower_sigma(self, sigma: float) -> [float]:
+        """
+        The lower error of every parameter marginalized in 1D at an input sigma value of its probability density
+        function (PDF), returned as a list.
+
+        See vector_at_sigma for a full description of how the parameters at sigma are computed.
+
+        Parameters
+        -----------
+        sigma
+            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
+        """
+        lowers = self.vector_at_lower_sigma(sigma=sigma)
+        return list(
+            map(
+                lambda lower, median_pdf: median_pdf - lower,
+                lowers,
+                self.median_pdf_vector,
+            )
+        )
+
+    def error_magnitude_vector_at_sigma(self, sigma: float) -> [float]:
+        """
+        The magnitude of every error after marginalization in 1D at an input sigma value of the probability density
+        function (PDF), returned as two lists of values corresponding to the lower and upper errors.
+
+        For example, if sigma is 1.0, the difference in the inferred values marginalized at 31.7% and 68.2% percentiles
+        of each PDF is returned.
+
+        Parameters
+        ----------
+        sigma
+            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
+        """
+        uppers = self.vector_at_upper_sigma(sigma=sigma)
+        lowers = self.vector_at_lower_sigma(sigma=sigma)
+        return list(map(lambda upper, lower: upper - lower, uppers, lowers))
+
+    def error_instance_at_sigma(self, sigma: float) -> ModelInstance:
+        """
+        The error of every parameter marginalized in 1D at an input sigma value of its probability density function
+        (PDF), returned as a list of model instances corresponding to the lower and upper errors.
+
+        See vector_at_sigma for a full description of how the parameters at sigma are computed.
+
+        Parameters
+        ----------
+        sigma
+            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
+        """
+        return self.model.instance_from_vector(
+            vector=self.error_magnitude_vector_at_sigma(sigma=sigma),
+            assert_priors_in_limits=False,
+        )
+
+    def error_instance_at_upper_sigma(self, sigma: float) -> ModelInstance:
+        """
+        The upper error of every parameter marginalized in 1D at an input sigma value of its probability density
+        function (PDF), returned as a model instance.
+
+        See vector_at_sigma for a full description of how the parameters at sigma are computed.
+
+        Parameters
+        -----------
+        sigma
+            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
+        """
+        return self.model.instance_from_vector(
+            vector=self.error_vector_at_upper_sigma(sigma=sigma),
+            assert_priors_in_limits=False,
+        )
+
+    def error_instance_at_lower_sigma(self, sigma: float) -> ModelInstance:
+        """
+        The lower error of every parameter marginalized in 1D at an input sigma value of its probability density
+        function (PDF), returned as a model instance.
+
+        See vector_at_sigma for a full description of how the parameters at sigma are computed.
+
+        Parameters
+        -----------
+        sigma
+            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
+        """
+        return self.model.instance_from_vector(
+            vector=self.error_vector_at_lower_sigma(sigma=sigma),
+            assert_priors_in_limits=False,
+        )
+
+    def gaussian_priors_at_sigma(self, sigma: float) -> [list]:
+        """
+        `GaussianPrior`s of every parameter used to link its inferred values and errors to priors used to sample the
+        same (or similar) parameters in a subsequent search, where:
+
+        - The mean is given by their most-probable values (using median_pdf_vector).
+        - Their errors are computed at an input sigma value (using errors_at_sigma).
+
+        Parameters
+        -----------
+        sigma
+            The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF).
+        """
+
+        means = self.median_pdf_vector
+
+        uppers = self.vector_at_upper_sigma(sigma=sigma)
+        lowers = self.vector_at_lower_sigma(sigma=sigma)
+
+        # noinspection PyArgumentList
+        sigmas = list(
+            map(
+                lambda mean, upper, lower: max([upper - mean, mean - lower]),
+                means,
+                uppers,
+                lowers,
+            )
+        )
+
+        return list(map(lambda mean, sigma: (mean, sigma), means, sigmas))
+
+    def log_likelihood_from_sample_index(self, sample_index: int) -> float:
+        """
+        The log likelihood of an individual sample of the non-linear search.
+
+        Parameters
+        ----------
+        sample_index
+            The index of the sample in the non-linear search, e.g. 0 gives the first sample.
+        """
+        raise NotImplementedError()
+
+    def vector_from_sample_index(self, sample_index: int) -> [float]:
+        """
+        The parameters of an individual sample of the non-linear search, returned as a 1D list.
+
+        Parameters
+        ----------
+        sample_index
+            The index of the sample in the non-linear search, e.g. 0 gives the first sample.
+        """
+        raise NotImplementedError()
+
+    def offset_vector_from_input_vector(self, input_vector: List) -> [float]:
+        """
+        The values of an input_vector offset by the median_pdf_vector (the PDF medians).
+
+        If the 'true' values of a model are known and input as the input_vector, this function returns the results
+        of the `NonLinearSearch` as values offset from the 'true' model. For example, a value 0.0 means the non-linear
+        search estimated the model parameter value correctly.
+
+        Parameters
+        ----------
+        input_vector
+            A 1D list of values the most-probable model is offset by. This list must have the same dimensions as free
+            parameters in the model-fit.
+        """
+        return list(
+            map(
+                lambda input, median_pdf: median_pdf - input,
+                input_vector,
+                self.median_pdf_vector,
+            )
+        )
+
+
+def quantile(x, q, weights=None):
+    """
+    Copied from corner.py
+
+    Compute sample quantiles with support for weighted samples.
+
+    Note
+    ----
+    When ``weight_list`` is ``None``, this method simply calls numpy's percentile
+    function with the values of ``q`` multiplied by 100.
+
+    Parameters
+    ----------
+    x : array_like[nsamples,]
+       The samples.
+    q : array_like[nquantiles,]
+       The list of quantiles to compute. These should all be in the range
+       ``[0, 1]``.
+    weights : Optional[array_like[nsamples,]]
+        An optional weight corresponding to each sample. These
+
+    Returns
+    -------
+    quantiles : array_like[nquantiles,]
+        The sample quantiles computed at ``q``.
+
+    Raises
+    ------
+    ValueError
+        For invalid quantiles; ``q`` not in ``[0, 1]`` or dimension mismatch between ``x`` and ``weight_list``.
+    """
+    x = np.atleast_1d(x)
+    q = np.atleast_1d(q)
+
+    if np.any(q < 0.0) or np.any(q > 1.0):
+        raise ValueError("Quantiles must be between 0 and 1")
+
+    if weights is None:
+        return np.percentile(x, list(100.0 * q))
+    else:
+        weights = np.atleast_1d(weights)
+        if len(x) != len(weights):
+            raise ValueError("Dimension mismatch: len(weight_list) != len(x)")
+        idx = np.argsort(x)
+        sw = weights[idx]
+        cdf = np.cumsum(sw)[:-1]
+        cdf /= cdf[-1]
+        cdf = np.append(0, cdf)
+        return np.interp(q, cdf, x[idx]).tolist()
